Add basics for incremental eval terms
